Apply a simple indentation format to the HTML code
Code indentation does not follow a unified format. There are child
elements with "less" indentation than their parents, tabs and spaces are
mixed, and a lot of lines have trailing spaces.

This commit applies a simple indentatino to whole the document: all the
subordinate blocks are indented with two spaces and all the tabs and
trailing are removed.