Add an "Arabic Script Overview" section
See #12 for the discussion. This section puts a nice introductory
overview on the beginning of the document and allows other parts of the
document to focus only on the issues that they want to talk about.

Some parts of the doucment, now covered here, are removed from other
sections. And now that the chapter "Characters" doesn't focus that much
on characters, it is renamed to "characters and words" to better reflect
its role.