Replace literal quotes with spans
This decision is discussed in part of #13. Using `span`s instead of
putting quotation marks in the text brings flexiblity to changing the
quotes easily and helps with localization of the document to other
languages.

Since these are the first times using these quotation spans, I found out
that the CSS is using straight quotation marks (' and "). I replaced
them with curly quotation marks (‘, ’, “, and ”).