Replace some of the characters with images
Some of the characters are not visually represented well enough with
their own character, like diacritics and control characters.
Representing those with images in the character charts would improve
our tables.

To support this, the Groovy script now checkws for those images and uses
them instead of character texts when avaialble.

The script sees in the direcotry "char-images" for files with names like
U+xxxx.svg. The "char-images" directory is a symlink to the document's
"images/characters" directory which is used directory in the HTML file.

This enables us to add image files with proper file names to the
"images/characters" directory and the script takes care of the rest.

Along with the screen update, this commit include SVG files for a bunch
of the character codes too.