Add some text for the diacritics section
This is the text that I had originally prepared for the "characters"
section, but no it's more fit for the "diacritics" section. This is by
no means what this section should look like. It's just a small note
about combining the diacritics that probably belongs here. The main work
on this section remains.