Note that justification mechanisms are not exclusive
Even though there is a title about combination of the justification
mechanisms at the end of the justification section, it might give a
false impression to the readers who are reading from top that the
mechanisms listed are not compatible with each other. This short note
should take care of that.