Use <h2> for titles everywhere
The actual header tag used does not matter, because ReSpec would replace
them with something between <h1> to <h6> based on the level of their
section, but for the sake of consistency it's better to use a single
header everywhere. ReSpec says <h2> is sort of a tacit convention, so I
use that.