Use ReSpec bibliography to manage external references
And update content to use the references (only where content is drafted,
and not where links are temporary/todo/tbd), alongside direct link to
the section/chapter (html/pdf) for easier use.

IMHO, the combination is easy to maintain using simple text search for
URLs, and the local bibliography db can act as a source for extenral
dependencies.

NOTE: I have to use the local setting because when working with
index.php on my browser using `file://` it cannot access the non-local
biblio db. (Please let me know if there's another way to fix this
problem.)

Fixes #103.