Use dfn/h5 instead of span/em/strong
Use `dfn` for term definitions, so we can link to them easily. Most of
the non-definition instances of `qterm` are replaced with links (`a`) to
the definitions. In some cases, like using terms from other document,
like Unicode resources, they are left alone as `qterm`. We can update
these instances later, as needed.

Also, in one case, `ol` and `strong` were being used to create
sub-sections with headers, which I replaced with `secion` and `h5`.

With this, there are no `strong` tags in use, except the list of TBD
items at the end. A few `em` cases still exist, which are not exactly
definitions. We can replace those with `qterm` if needed.